fix setup.py

+++ setup.py
@@ -10,7 +10,8 @@ binaries = {
     'imagick-6.7.2.10-win32.zip': '1a9e7c9514fde117737de4f7e730db40'
 }
 
-def md5_file(f, block_size=1**20):
+
+def md5_file(f, block_size=1 ** 20):
     f = open(f, 'rb')
     fingerprint = md5()
     while True:
@@ -22,6 +23,7 @@ def md5_file(f, block_size=1**20):
     
     return fingerprint.hexdigest()
 
+
 def extract(zipfile, path):
     z = ZipFile(zipfile)
     for name in z.namelist():
@@ -29,6 +31,7 @@ def extract(zipfile, path):
     z.close()
 
 from distutils.command.build import build
+
 
 class tinyimg_build(build):
     def run(self):
@@ -71,9 +74,12 @@ class tinyimg_build(build):
         return result
 
 
-from os.path import join 
+from os.path import join, exists
 from hashlib import md5
-from urllib import urlretrieve
+try:
+    from urllib import urlretrieve
+except ImportError:
+    from urllib.request import urlretrieve  # @Reimport
 from zipfile import ZipFile
 
 from distutils.dir_util import mkpath
@@ -84,26 +90,25 @@ try:
     from setuptools.command.install import install
 except ImportError:
     # still it can work with python setup.py install
-    from distutils.command.install import install
+    from distutils.command.install import install  # @Reimport
+
 
 class tinyimg_install(install):
     def run(self):
-        from os.path import join 
-        
         result = install.run(self)
         
         lib_base = join(self.build_base, 'libraries')
-        self.copy_tree(lib_base, join(self.install_lib, 'tinyimg/lib'))
+        if exists(lib_base):
+            self.copy_tree(lib_base, join(self.install_lib, 'tinyimg/lib'))
         
         return result
 
-install_requires=['six', 'decorator']
+install_requires = ['six', 'decorator']
 
 try:
     format
 except NameError:
     install_requires.append('StringFormat')
-
 
 from setuptools import setup
 
@@ -111,7 +116,7 @@ setup(
     name='tinyimg',
     description='Python raster imaging library',
     author='Paweł Piotr Przeradowski',
-    author_email = 'przeradowski@gmail.com',
+    author_email='przeradowski@gmail.com',
     url='https://bitbucket.org/squeaky/tinyimg',
     version='0.1dev',
     packages=['tinyimg', 'tinyimg.api', 'tinyimg.tests', 'tinyimg.api.tests'],
